added shared warlock abilities.ts and removed them for the spec specific abilities
