+++ src/analysis/retail/warlock/affliction/modules/features/Abilities.ts
@@ -1,11 +1,12 @@
 import SPELLS from 'common/SPELLS';
 import TALENTS from 'common/TALENTS/warlock';
 import ISSUE_IMPORTANCE from 'parser/core/ISSUE_IMPORTANCE';
-import CoreAbilities from 'parser/core/modules/Abilities';
+import SharedAbilities from 'analysis/retail/warlock/shared/Abilities';
 import SPELL_CATEGORY from 'parser/core/SPELL_CATEGORY';
+import { SpellbookAbility } from 'parser/core/modules/Ability';
 
-class Abilities extends CoreAbilities {
-  spellbook() {
+class Abilities extends SharedAbilities {
+  spellbook(): SpellbookAbility[] {
     const combatant = this.selectedCombatant;
     return [
       // Rotational
@@ -159,21 +160,6 @@ class Abilities extends CoreAbilities {
           majorIssueEfficiency: 0.1,
         },
       },
-      {
-        spell: TALENTS.DARK_PACT_TALENT.id,
-        buffSpellId: TALENTS.DARK_PACT_TALENT.id,
-        category: SPELL_CATEGORY.DEFENSIVE,
-        cooldown: combatant.hasTalent(TALENTS.FREQUENT_DONOR_TALENT) ? 45 : 60,
-        gcd: null,
-        enabled: combatant.hasTalent(TALENTS.DARK_PACT_TALENT),
-        castEfficiency: {
-          suggestion: true,
-          importance: ISSUE_IMPORTANCE.MINOR,
-          recommendedEfficiency: 0.33,
-          averageIssueEfficiency: 0.2,
-          majorIssueEfficiency: 0.1,
-        },
-      },
       // Utility
       {
         spell: TALENTS.BURNING_RUSH_TALENT.id,
@@ -198,28 +184,6 @@ class Abilities extends CoreAbilities {
         enabled: combatant.hasTalent(TALENTS.MORTAL_COIL_TALENT),
         gcd: {
           base: 1500,
-        },
-      },
-      {
-        spell: SPELLS.DEMONIC_CIRCLE_SUMMON.id,
-        category: SPELL_CATEGORY.UTILITY,
-        gcd: {
-          base: 1500,
-        },
-        cooldown: 10,
-        castEfficiency: {
-          suggestion: false,
-        },
-      },
-      {
-        spell: SPELLS.DEMONIC_CIRCLE_TELEPORT.id,
-        category: SPELL_CATEGORY.UTILITY,
-        cooldown: 30,
-        gcd: {
-          base: 1500,
-        },
-        castEfficiency: {
-          suggestion: false,
         },
       },
       {
@@ -367,6 +331,7 @@ class Abilities extends CoreAbilities {
         cooldown: 15,
         gcd: null,
       },
+      ...super.spellbook(),
     ];
   }
 }
